requirements.txt
bidict==0.21.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
eventlet==0.33.0
Flask==2.0.2
gevent==21.8.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
zope.event==4.5.0
zope.interface==5.4.0
